Exception handling - DELETE

+++ src/main/java/org/springbootapi/api/services/UserService.java
@@ -3,7 +3,11 @@ package org.springbootapi.api.services;
 
 import org.springbootapi.api.entities.User;
 import org.springbootapi.api.repositories.UserRepository;
+import org.springbootapi.api.services.exceptions.DatabaseException;
+import org.springbootapi.api.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -32,7 +36,13 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        try {
+            if (!repository.existsById(id))
+                throw new ResourceNotFoundException(id);
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User user) {
